Fix Debugger's break.
The break wasn't stopping in the instruction at the address provided.
Currently, it runs the instructions until the one in the address
provided and runs that instruction as well.

I don't believe currently there is an easy way to stop the code just
before executing the instruction at the provided address. The best way
to do this, would to create the instruction object before running its
code, however it doesn't look nice, because there would be some kind of
duplicate code (i.e., probably some `match` with the same arms).