Fix debugger break.
If the user types an address that doesn't correspond to the start of a
instruction, the debugger will be in a loop forever.

It *has* to be `instruction.address == addr` because some jumps can go
past the address and then it can come back to it. So, it will stop
correctly, even if the PC is pointing to some instruction after address.