Use a different approach for rendering a frame.
Instead of rendering each line, now we hold a buffer of the size of the screen and update each
line at the correct moment (HBlank event) and the display is rendered only after the
VBlank event.

This is for trying to keep things more in sync and looks like it gave a speed boost
to the "nintendo logo" animation.